move unsigned casting to after column parsing

+++ src/store/accounts.rs
@@ -97,13 +97,13 @@ impl Store {
             .map_err(StoreError::QueryError)?
             .query_map([], |row| {
                 let row : i64 = row.get(0)?;
-                Ok(row as u64)
+                Ok(row)
             })
             .expect("no binding parameters used in query")
             .map(|result| {
                 result
                     .map_err(StoreError::ColumnParsingError)
-                    .map(|id: u64| AccountId::try_from(id).expect("account id is not valid"))
+                    .map(|id: i64| AccountId::try_from(id as u64).expect("account id is not valid"))
             })
             .collect::<Result<Vec<AccountId>, _>>()
     }
